Split resource tests
