+++ frictionless/resource.py
@@ -1118,7 +1118,7 @@ class Resource(Metadata):
         parser.write_row_stream(self.to_copy())
         return target
 
-    # Import/Export
+    # Export/Import
 
     def to_dict(self):
         """Create a dict from the resource
