+++ tests/resource/test_onerror.py
@@ -1,0 +1,51 @@
+import pytest
+from frictionless import Resource, FrictionlessException, helpers
+
+
+IS_UNIX = not helpers.is_platform("windows")
+BASEURL = "https://raw.githubusercontent.com/frictionlessdata/frictionless-py/master/%s"
+
+
+# Onerror
+
+
+def test_resource_onerror():
+    resource = Resource(path="data/invalid.csv")
+    assert resource.onerror == "ignore"
+    assert resource.read_rows()
+
+
+def test_resource_onerror_header_warn():
+    data = [["name"], [1], [2], [3]]
+    schema = {"fields": [{"name": "bad", "type": "integer"}]}
+    resource = Resource(data=data, schema=schema, onerror="warn")
+    assert resource.onerror == "warn"
+    with pytest.warns(UserWarning):
+        resource.read_rows()
+
+
+def test_resource_onerror_header_raise():
+    data = [["name"], [1], [2], [3]]
+    schema = {"fields": [{"name": "bad", "type": "integer"}]}
+    resource = Resource(data=data, schema=schema, onerror="raise")
+    assert resource.onerror == "raise"
+    with pytest.raises(FrictionlessException):
+        resource.read_rows()
+
+
+def test_resource_onerror_row_warn():
+    data = [["name"], [1], [2], [3]]
+    schema = {"fields": [{"name": "name", "type": "string"}]}
+    resource = Resource(data=data, schema=schema, onerror="warn")
+    assert resource.onerror == "warn"
+    with pytest.warns(UserWarning):
+        resource.read_rows()
+
+
+def test_resource_onerror_row_raise():
+    data = [["name"], [1], [2], [3]]
+    schema = {"fields": [{"name": "name", "type": "string"}]}
+    resource = Resource(data=data, schema=schema, onerror="raise")
+    assert resource.onerror == "raise"
+    with pytest.raises(FrictionlessException):
+        resource.read_rows()
